refactor(tui): hint constants own their view modules (#309)
Previous commit d4a9d40 left definitions in hint-map.ts with re-exports
from view modules. Spec requires the inverse: each view DEFINES its
KeyAction[] const; hint-map IMPORTS them. Fixes screen-manager.test.ts
mocks to expose the new constants so the assembled hint-map can
resolve them through the mocks.
