+++ src/tui/screens/goal-input.tsx
@@ -151,4 +151,10 @@ export const GoalInput: React.NamedExoticComponent<GoalInputProps> = React.memo(
   );
 });
 
-export { GOAL_INPUT_HINTS } from "../data/hint-map.js";
+import type { KeyAction } from "../data/hint-map.js";
+
+export const GOAL_INPUT_HINTS: readonly KeyAction[] = Object.freeze([
+  { key: "Enter", label: "Continue" },
+  { key: "Esc", label: "Back" },
+  { key: "Ctrl+U", label: "Clear" },
+]);
