+++ src/tui/screens/running-view.tsx
@@ -1826,4 +1826,13 @@ function renderStatusBar(
   );
 }
 
-export { RUNNING_VIEW_HINTS } from "../data/hint-map.js";
+import type { KeyAction } from "../data/hint-map.js";
+
+export const RUNNING_VIEW_HINTS: readonly KeyAction[] = Object.freeze([
+  { key: ":", label: "Goto" },
+  { key: "/", label: "Filter" },
+  { key: "1-5", label: "Panel" },
+  { key: "Esc", label: "Back" },
+  { key: "?", label: "Help" },
+  { key: "q", label: "Quit" },
+]);
